+++ classes/class-wicket-acc-assets.php
@@ -42,6 +42,17 @@ class Assets extends WicketAcc
      */
     public function enqueue_frontend_assets()
     {
+        global $post;
+
+        // Only on pages related to Wicket Account Centre
+        if (!($post->post_type === 'my-account' ||
+        str_starts_with($post->post_name, 'organization') ||
+        is_woocommerce() ||
+            is_account_page() ||
+            is_wc_endpoint_url())) {
+            return;
+        }
+
         wp_enqueue_style('wicket-acc-frontend-styles', WICKET_ACC_URL . 'assets/css/wicket-acc-main.css', [], WICKET_ACC_VERSION);
         wp_enqueue_script('wicket-acc-frontend-scripts', WICKET_ACC_URL . 'assets/js/wicket-acc-main.js', [], WICKET_ACC_VERSION, true);
         wp_enqueue_script('wicket-acc-frontend-legacy-scripts', WICKET_ACC_URL . 'assets/js/wicket-acc-legacy.js', [], WICKET_ACC_VERSION, true);